Move code to the appropriate places.   Adjust CreateAsyncThreadPool interface
