--- async/random_read_context.cc
@@ -1,13 +1,0 @@
-//  Copyright (c) 2011-present, Facebook, Inc.  All rights reserved.
-//  This source code is licensed under the BSD-style license found in the
-//  LICENSE file in the root directory of this source tree. An additional grant
-//  of patent rights can be found in the PATENTS file in the same directory.
-//
-
-#include "async/random_read_context.h"
-
-namespace rocksdb {
-namespace async {
-
-} // namespace async
-} // namespace rocksdb
